+++ CHANGES.rst
@@ -22,6 +22,7 @@ Minor Changes
 
 - build(signer): include socket time out error in catch statements
 - docs(fix): unlink docstring reference to changed parameter name
+- fix(config): allow setting the time unit for passcrypt instances
 - fix(db): remove deprecated use of dunder dict on containers
 - refactor(config): simplify interface for low-level Passcrypt helpers